Update functional tests to pass <PD-1471>
